feat(trusty-mpm): add `[hooks] prompt_context` opt-out for the per-prompt memory injection
The `UserPromptSubmit` -> `trusty-memory prompt-context` hook injects a
measured ~1,211 tokens into every prompt (median 1,252, range 693-1,438
across 1,114 firings over five days) — roughly 24,000 tokens across a
20-turn session, recurring. #4904 measured 0 clean matches out of 17
curated facts on that same corpus.

There was no way to stop paying it. The hook block is a hardcoded const
written unconditionally at every session launch, so a hand edit to
`.claude/settings.json` was overwritten on the next launch.

`[hooks] prompt_context = false` in `~/.trusty-mpm/config.toml` now
suppresses it, read through the existing `core::config` reader. Default
is `true`: an absent section, and a present-but-empty one, both leave the
write byte-identical to before (verified by diffing the generated
settings.json against origin/main — same md5). Only the
`UserPromptSubmit` entry is affected; `SessionStart`, the PM guard, and
the six-event lifecycle triad are written either way. Config file only —
no CLI flag, no environment variable.

The strip that removes trusty-mpm's own prior entries now covers every
event trusty-mpm owns rather than only the ones the current config
writes. Without that the key would have done nothing on any project
already launched once: the stale `UserPromptSubmit` entry would have
stayed in `.claude/settings.json` and kept firing.

`config.rs`'s inline test module moves to a sibling `config_tests.rs` —
the new section pushed the production file to 533 SLOC, over the 500 cap.
No test changed in the move.

Closes #5034
