+++ crates/trusty-mpm/src/core/session_launch/project_hooks_tests.rs
@@ -18,7 +18,7 @@ use tempfile::TempDir;
 
 #[test]
 fn project_managed_hook_additions_combines_all_three_sources() {
-    let additions = project_managed_hook_additions();
+    let additions = project_managed_hook_additions(true);
     let hooks = additions["hooks"]
         .as_object()
         .expect("hooks must be an object");
@@ -71,9 +71,76 @@ fn project_managed_hook_additions_combines_all_three_sources() {
 fn project_managed_hook_additions_is_stable_across_calls() {
     // Why: two independent calls must produce byte-identical output so the
     // caller's merge is idempotent across repeated `write_project_hooks` runs.
-    let first = project_managed_hook_additions();
-    let second = project_managed_hook_additions();
+    let first = project_managed_hook_additions(true);
+    let second = project_managed_hook_additions(true);
     assert_eq!(first, second);
+}
+
+/// Why (#5034): `[hooks] prompt_context = false` must drop the
+/// `UserPromptSubmit` entry and NOTHING else. The regression this guards is a
+/// toggle that also takes out `SessionStart` — the other `TRUSTY_MEMORY_HOOKS`
+/// key, sharing the same constant and the same `trusty-memory ` command prefix.
+/// What: builds the additions with the toggle off and asserts `UserPromptSubmit`
+/// is absent while `SessionStart`, the PM guard, and all six lifecycle-triad
+/// events are byte-identical to the enabled build.
+#[test]
+fn project_managed_hook_additions_omits_prompt_context_when_disabled() {
+    let enabled = project_managed_hook_additions(true);
+    let disabled = project_managed_hook_additions(false);
+
+    let off = disabled["hooks"].as_object().expect("hooks is an object");
+    assert!(
+        !off.contains_key("UserPromptSubmit"),
+        "UserPromptSubmit must be absent when the toggle is off: {off:?}"
+    );
+
+    // Every other event must survive with its enabled-build value untouched.
+    let on = enabled["hooks"].as_object().unwrap();
+    for event in [
+        "SessionStart",
+        "PreToolUse",
+        "PostToolUse",
+        "Stop",
+        "SubagentStop",
+        "SessionEnd",
+    ] {
+        assert_eq!(
+            off.get(event),
+            on.get(event),
+            "{event} must be unchanged by the prompt_context toggle"
+        );
+    }
+    assert_eq!(
+        off.len(),
+        on.len() - 1,
+        "exactly one event key may differ between the two builds"
+    );
+}
+
+/// Why (#5034): the strip domain must not shrink with the toggle, or the
+/// opt-out silently does nothing on any project already launched once.
+/// What: asserts the owned-event list covers the key set of BOTH toggle
+/// variants.
+#[test]
+fn project_managed_hook_events_is_a_superset_of_every_variant() {
+    let owned = project_managed_hook_events();
+    for enabled in [true, false] {
+        for key in project_managed_hook_additions(enabled)["hooks"]
+            .as_object()
+            .unwrap()
+            .keys()
+        {
+            assert!(
+                owned.contains(key),
+                "{key} (toggle={enabled}) must be in the owned strip domain: {owned:?}"
+            );
+        }
+    }
+    assert!(
+        owned.iter().any(|e| e == "UserPromptSubmit"),
+        "UserPromptSubmit must stay in the strip domain even though the \
+         disabled build never writes it: {owned:?}"
+    );
 }
 
 #[test]
@@ -109,7 +176,7 @@ fn write_project_hooks_writes_lifecycle_triad() {
     let tmp = TempDir::new().unwrap();
     let project = tmp.path();
 
-    super::super::settings::write_project_hooks(project).expect("write succeeds");
+    super::super::settings::write_project_hooks(project, true).expect("write succeeds");
 
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
@@ -164,7 +231,7 @@ fn write_project_hooks_preserves_foreign_hooks() {
     )
     .unwrap();
 
-    super::super::settings::write_project_hooks(project).expect("write succeeds");
+    super::super::settings::write_project_hooks(project, true).expect("write succeeds");
 
     let value: serde_json::Value =
         serde_json::from_str(&std::fs::read_to_string(claude_dir.join("settings.json")).unwrap())
@@ -186,7 +253,7 @@ fn write_project_hooks_preserves_foreign_hooks() {
     );
 
     // Re-running must not duplicate the foreign entry or our own groups.
-    super::super::settings::write_project_hooks(project).expect("second write succeeds");
+    super::super::settings::write_project_hooks(project, true).expect("second write succeeds");
     let value2: serde_json::Value =
         serde_json::from_str(&std::fs::read_to_string(claude_dir.join("settings.json")).unwrap())
             .unwrap();
@@ -221,7 +288,7 @@ fn write_project_hooks_writes_via_atomic_path() {
     let bak_path = claude_dir.join("settings.json.bak");
     let tmp_path = claude_dir.join("settings.json.tmp");
 
-    super::super::settings::write_project_hooks(project).expect("first write succeeds");
+    super::super::settings::write_project_hooks(project, true).expect("first write succeeds");
     assert!(settings_path.exists(), "settings.json must be created");
     assert!(
         !bak_path.exists(),
@@ -233,7 +300,7 @@ fn write_project_hooks_writes_via_atomic_path() {
     );
     let first_content = std::fs::read_to_string(&settings_path).unwrap();
 
-    super::super::settings::write_project_hooks(project).expect("second write succeeds");
+    super::super::settings::write_project_hooks(project, true).expect("second write succeeds");
     assert!(
         bak_path.exists(),
         "write_json_atomic must back up the prior file before replacing it"
@@ -251,4 +318,210 @@ fn write_project_hooks_writes_via_atomic_path() {
     let value: serde_json::Value =
         serde_json::from_str(&std::fs::read_to_string(&settings_path).unwrap()).unwrap();
     assert!(value["hooks"]["SessionStart"].is_array());
+}
+
+/// Why (#5034): the end-to-end opt-out — `[hooks] prompt_context = false` must
+/// leave no `UserPromptSubmit` entry in the file the launched session reads,
+/// while every other hook trusty-mpm writes is still there. The measured cost
+/// this buys back is ~1,211 tokens per prompt (#4904).
+/// What: writes with the toggle off into a fresh project and asserts
+/// `UserPromptSubmit` is absent, `SessionStart` still carries the
+/// `trusty-memory inbox-check` group, `PreToolUse` still carries the PM guard,
+/// and all six lifecycle-triad events are present.
+#[test]
+fn write_project_hooks_omits_prompt_context_when_disabled() {
+    let tmp = TempDir::new().unwrap();
+    let project = tmp.path();
+
+    super::super::settings::write_project_hooks(project, false).expect("write succeeds");
+
+    let value: serde_json::Value = serde_json::from_str(
+        &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
+    )
+    .unwrap();
+    let hooks = value["hooks"].as_object().expect("hooks must be an object");
+
+    assert!(
+        !hooks.contains_key("UserPromptSubmit"),
+        "UserPromptSubmit must not be written when disabled: {hooks:?}"
+    );
+    let raw = serde_json::to_string(&value).unwrap();
+    assert!(
+        !raw.contains("prompt-context"),
+        "no prompt-context command may survive anywhere in the file: {raw}"
+    );
+
+    // The SessionStart trusty-memory group is a separate hook and must stay.
+    let session_start = hooks["SessionStart"].as_array().unwrap();
+    let ss_commands: Vec<&str> = session_start
+        .iter()
+        .map(|g| g["hooks"][0]["command"].as_str().unwrap())
+        .collect();
+    assert!(
+        ss_commands.contains(&"trusty-memory inbox-check"),
+        "SessionStart must keep the inbox-check hook: {ss_commands:?}"
+    );
+
+    // The PM guard and the full lifecycle triad must be unaffected.
+    let pre_commands: Vec<&str> = hooks["PreToolUse"]
+        .as_array()
+        .unwrap()
+        .iter()
+        .map(|g| g["hooks"][0]["command"].as_str().unwrap())
+        .collect();
+    assert!(
+        pre_commands.iter().any(|c| c.ends_with(" hook --pm-guard")),
+        "the PM guard must still be registered: {pre_commands:?}"
+    );
+    for event in [
+        "PreToolUse",
+        "PostToolUse",
+        "Stop",
+        "SubagentStop",
+        "SessionStart",
+        "SessionEnd",
+    ] {
+        assert!(
+            hooks[event]
+                .as_array()
+                .unwrap()
+                .iter()
+                .any(|g| g["hooks"][0]["command"]
+                    .as_str()
+                    .is_some_and(|c| c.ends_with(" hook"))),
+            "{event} must still carry the lifecycle-triad command"
+        );
+    }
+}
+
+/// Why (#5034): this is the test that fails if the strip domain is derived from
+/// the additions instead of the owned set. Every project in the wild has
+/// already been launched with the hook enabled, so the entry is ALREADY in
+/// `.claude/settings.json`; an opt-out that only stops re-writing it would
+/// leave it firing forever and the config key would appear to do nothing.
+/// What: writes once enabled (seeding the entry), then writes disabled, and
+/// asserts the stale entry is gone while a foreign `UserPromptSubmit` entry
+/// added by the operator survives.
+#[test]
+fn write_project_hooks_strips_stale_prompt_context_when_disabled() {
+    let tmp = TempDir::new().unwrap();
+    let project = tmp.path();
+    let settings_path = project.join(".claude").join("settings.json");
+
+    super::super::settings::write_project_hooks(project, true).expect("enabled write succeeds");
+    let seeded: serde_json::Value =
+        serde_json::from_str(&std::fs::read_to_string(&settings_path).unwrap()).unwrap();
+    assert_eq!(
+        seeded["hooks"]["UserPromptSubmit"][0]["hooks"][0]["command"],
+        serde_json::json!("trusty-memory prompt-context"),
+        "precondition: the enabled write must have seeded the entry"
+    );
+
+    // The operator also has a hook of their own on the same event.
+    let mut with_foreign = seeded.clone();
+    with_foreign["hooks"]["UserPromptSubmit"]
+        .as_array_mut()
+        .unwrap()
+        .push(serde_json::json!({
+            "matcher": "",
+            "hooks": [{ "type": "command", "command": "my-own-tool inject", "timeout": 5 }]
+        }));
+    std::fs::write(
+        &settings_path,
+        serde_json::to_string_pretty(&with_foreign).unwrap(),
+    )
+    .unwrap();
+
+    super::super::settings::write_project_hooks(project, false).expect("disabled write succeeds");
+
+    let value: serde_json::Value =
+        serde_json::from_str(&std::fs::read_to_string(&settings_path).unwrap()).unwrap();
+    let raw = serde_json::to_string(&value).unwrap();
+    assert!(
+        !raw.contains("trusty-memory prompt-context"),
+        "the stale entry from the earlier enabled launch must be stripped: {raw}"
+    );
+
+    let remaining: Vec<&str> = value["hooks"]["UserPromptSubmit"]
+        .as_array()
+        .expect("the operator's own group keeps the event key alive")
+        .iter()
+        .map(|g| g["hooks"][0]["command"].as_str().unwrap())
+        .collect();
+    assert_eq!(
+        remaining,
+        vec!["my-own-tool inject"],
+        "a foreign UserPromptSubmit entry must survive the strip"
+    );
+}
+
+/// Why (#5034): the opt-out must be reversible — clearing the config key has to
+/// restore the hook, not leave the project permanently stripped.
+/// What: enabled → disabled → enabled, asserting the final file is
+/// byte-identical to the first enabled write.
+#[test]
+fn write_project_hooks_re_enabling_restores_the_hook() {
+    let tmp = TempDir::new().unwrap();
+    let project = tmp.path();
+    let settings_path = project.join(".claude").join("settings.json");
+
+    super::super::settings::write_project_hooks(project, true).expect("write 1");
+    let first = std::fs::read_to_string(&settings_path).unwrap();
+    super::super::settings::write_project_hooks(project, false).expect("write 2");
+    super::super::settings::write_project_hooks(project, true).expect("write 3");
+    let third = std::fs::read_to_string(&settings_path).unwrap();
+
+    assert_eq!(
+        third, first,
+        "re-enabling must reproduce the enabled write byte for byte"
+    );
+}
+
+/// Why (#5034): the default path must not move. `#[hooks] prompt_context`
+/// defaults to `true`, and the enabled write must be exactly what the
+/// pre-toggle code produced — which, for the strip half of the change, means
+/// the widened strip domain has to be a no-op when the hook is enabled.
+/// What: asserts the config default is `true`, and that the enabled write is
+/// unchanged whether the strip runs over the owned event set or over the
+/// additions' own keys (the pre-#5034 derivation), on both a fresh project and
+/// one already carrying a prior write.
+#[test]
+fn write_project_hooks_enabled_output_is_unchanged_by_the_toggle() {
+    assert!(
+        crate::core::config::MpmConfig::default()
+            .hooks
+            .prompt_context,
+        "the shipped default must keep the hook enabled"
+    );
+
+    // The pre-#5034 strip domain was the additions' own key set. With the hook
+    // enabled the two derivations must be identical, which is what makes the
+    // default path byte-identical to before.
+    let mut from_additions: Vec<String> = project_managed_hook_additions(true)["hooks"]
+        .as_object()
+        .unwrap()
+        .keys()
+        .cloned()
+        .collect();
+    let mut owned = project_managed_hook_events();
+    from_additions.sort();
+    owned.sort();
+    assert_eq!(
+        owned, from_additions,
+        "with the hook enabled the widened strip domain must be a no-op"
+    );
+
+    // And the write itself is stable across repeats, on a project that already
+    // carries a prior launch's entries.
+    let tmp = TempDir::new().unwrap();
+    let project = tmp.path();
+    let settings_path = project.join(".claude").join("settings.json");
+    super::super::settings::write_project_hooks(project, true).expect("write 1");
+    let first = std::fs::read_to_string(&settings_path).unwrap();
+    super::super::settings::write_project_hooks(project, true).expect("write 2");
+    assert_eq!(
+        std::fs::read_to_string(&settings_path).unwrap(),
+        first,
+        "the enabled write must be idempotent"
+    );
 }
