+++ crates/trusty-mpm/src/core/session_launch/tests.rs
@@ -995,7 +995,7 @@ fn write_project_hooks_writes_all_event_types() {
     let tmp = tempdir().unwrap();
     let project = tmp.path();
 
-    write_project_hooks(project).expect("write succeeds");
+    write_project_hooks(project, true).expect("write succeeds");
 
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
@@ -1025,7 +1025,7 @@ fn write_project_hooks_uses_canonical_commands() {
     let tmp = tempdir().unwrap();
     let project = tmp.path();
 
-    write_project_hooks(project).expect("write succeeds");
+    write_project_hooks(project, true).expect("write succeeds");
 
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
@@ -1059,7 +1059,7 @@ fn write_project_hooks_omits_post_tool_use_and_stop() {
     let tmp = tempdir().unwrap();
     let project = tmp.path();
 
-    write_project_hooks(project).expect("write succeeds");
+    write_project_hooks(project, true).expect("write succeeds");
 
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
@@ -1089,7 +1089,7 @@ fn write_project_hooks_registers_pm_guard() {
     let tmp = tempdir().unwrap();
     let project = tmp.path();
 
-    write_project_hooks(project).expect("write succeeds");
+    write_project_hooks(project, true).expect("write succeeds");
 
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
@@ -1130,8 +1130,8 @@ fn write_project_hooks_replaces_existing() {
     let tmp = tempdir().unwrap();
     let project = tmp.path();
 
-    write_project_hooks(project).expect("first write succeeds");
-    write_project_hooks(project).expect("second write succeeds");
+    write_project_hooks(project, true).expect("first write succeeds");
+    write_project_hooks(project, true).expect("second write succeeds");
 
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
@@ -1146,7 +1146,7 @@ fn write_project_hooks_replaces_existing() {
         "re-running must not duplicate our own handler groups"
     );
     // Unrelated keys must survive the replace.
-    write_project_hooks(project).expect("third write succeeds");
+    write_project_hooks(project, true).expect("third write succeeds");
     let value: serde_json::Value = serde_json::from_str(
         &std::fs::read_to_string(project.join(".claude").join("settings.json")).unwrap(),
     )
